feat: Created MusicClient
