+++ androidApp/src/main/java/com/devjeong/watermelon_player/android/music/presentation/AndroidMusicViewModel.kt
@@ -1,0 +1,4 @@
+package com.devjeong.watermelon_player.android.music.presentation
+
+class AndroidMusicViewModel {
+}
