+++ shared/src/commonMain/kotlin/com/devjeong/watermelon_player/audioplayer/data/Music/KtorMusicClient.kt
@@ -1,0 +1,44 @@
+package com.devjeong.watermelon_player.audioplayer.data.Music
+
+import NetworkConstants
+import io.ktor.client.*
+import io.ktor.client.engine.cio.*
+import io.ktor.client.plugins.*
+import io.ktor.client.plugins.contentnegotiation.*
+import io.ktor.serialization.kotlinx.json.*
+import io.ktor.util.logging.Logger
+import kotlinx.serialization.json.Json
+
+object KtorMusicClient {
+
+    const val TAG = "API 체크"
+
+    private val json = Json {
+        encodeDefaults = true
+        ignoreUnknownKeys = true
+    }
+
+    val httpClient = HttpClient(CIO) {
+        install(ContentNegotiation) {
+            json(json)
+        }
+
+        install(HttpTimeout) {
+            requestTimeoutMillis = 15_000
+            connectTimeoutMillis = 15_000
+            socketTimeoutMillis = 15_000
+        }
+
+        defaultRequest {
+            headers.append("Accept", "application/json")
+            headers.append("Content-Type", "application/json")
+            headers.append("apikey", NetworkConstants.supabase.supabaseUrl)
+        }
+
+    }
+}
+
+
+
+
+
