+++ shared/src/commonMain/kotlin/com/devjeong/watermelon_player/audioplayer/data/Music/MusicDto.kt
@@ -1,9 +1,9 @@
-package com.devjeong.watermelon_player.audioplayer.domain
+package com.devjeong.watermelon_player.audioplayer.data.Music
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Music(
+data class MusicDto(
     val streamUrl: String,
     val imageUrl: String,
     val artists: String,
