+++ shared/src/commonMain/kotlin/com/devjeong/watermelon_player/audioplayer/domain/music/MusicClient.kt
@@ -1,0 +1,7 @@
+package com.devjeong.watermelon_player.audioplayer.domain.music
+
+import com.devjeong.watermelon_player.audioplayer.data.Music.MusicDto
+
+interface MusicClient {
+    suspend fun getMusicList(): List<MusicDto>
+}
